BoD Use Case/requirements.txt
# BoD Presentation Analysis POC - Requirements

# Core Python dependencies
streamlit>=1.28.0
python-dotenv>=1.0.0
python-multipart>=0.0.6

# Document processing
PyMuPDF>=1.23.0          # PDF processing
python-pptx>=0.6.21      # PowerPoint processing
pytesseract>=0.3.10      # OCR integration
Pillow>=10.0.0           # Image processing

# AI and ML
openai>=1.3.0            # OpenAI API
mistralai>=0.0.8         # Mistral AI API
sentence-transformers>=2.2.2  # Local embeddings
chromadb>=0.4.15         # Vector database
ollama>=0.1.7            # Local LLM integration

# Data processing and analysis
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
nltk>=3.8.1
spacy>=3.7.0

# Text processing and NLP
regex>=2023.8.8
beautifulsoup4>=4.12.2
fuzzywuzzy>=0.18.0
python-Levenshtein>=0.21.1

# Utilities
requests>=2.31.0
aiohttp>=3.8.5
tqdm>=4.66.0
colorama>=0.4.6
loguru>=0.7.2

# Development and testing
pytest>=7.4.0
pytest-asyncio>=0.21.1
black>=23.7.0
flake8>=6.0.0

# Export and reporting
reportlab>=4.0.4         # PDF generation
jinja2>=3.1.2           # Template rendering
markdown>=3.5.1         # Markdown processing

# File handling
pathlib2>=2.3.7
watchdog>=3.0.0

# Optional: For advanced OCR (if Tesseract insufficient)
# easyocr>=1.7.0         # Alternative OCR engine
# paddleocr>=2.7.0       # Alternative OCR engine

# System dependencies (install separately):
# - Tesseract OCR: https://github.com/UB-Mannheim/tesseract/wiki
# - Ollama: https://ollama.ai/download
# - spaCy language model: python -m spacy download en_core_web_sm